complete all

+++ src/main/java/com/sna/homework003/service/impl/EventServiceImpl.java
@@ -3,15 +3,12 @@ package com.sna.homework003.service.impl;
 import com.sna.homework003.exception.DuplicateException;
 import com.sna.homework003.exception.NotFoundException;
 import com.sna.homework003.model.dto.EventDTO;
-import com.sna.homework003.model.entity.Attendees;
 import com.sna.homework003.model.entity.Events;
-import com.sna.homework003.repository.AttendeeRepository;
 import com.sna.homework003.repository.EventAttendeeRepository;
 import com.sna.homework003.repository.EventRepository;
 import com.sna.homework003.service.AttendeeService;
 import com.sna.homework003.service.EventService;
 import com.sna.homework003.service.VenueService;
-import jdk.jfr.Event;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -24,8 +21,8 @@ public class EventServiceImpl implements EventService {
     private final AttendeeService attendeeService;
     private final VenueService venueService;
     @Override
-    public List<Events> getAllEvents() {
-        return eventRepository.getAllEvents();
+    public List<Events> getAllEvents(Integer size, Integer page) {
+        return eventRepository.getAllEvents(size,page);
     }
 
     @Override
@@ -34,12 +31,10 @@ public class EventServiceImpl implements EventService {
         for (Integer attendeeId : request.getAttendeeId()) {
             attendeeService.getAttendeeById(attendeeId);
         }
-
-        Events existing = eventRepository.getEventByNameAndDate(request.getEventName(), request.getEventDate());
+        Events existing = eventRepository.getAllEventByNameAndDate(request.getEventName(), request.getEventDate());
         if (existing != null) {
             throw new DuplicateException("Event with name " + request.getEventName() + " and date " + request.getEventDate() + " already exists");
         }
-
         Events event = eventRepository.saveEvent(request);
         for (Integer attendeeId : request.getAttendeeId()) {
             eventAttendeeRepository.saveEventAttendee(event.getEventId(), attendeeId);
@@ -59,16 +54,30 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public Events updateEvent(Integer eventId, EventDTO request) {
-        getEventById(eventId);
         venueService.getVenueById(request.getVenueId());
-        Events event=  eventRepository.updateEvent(eventId, request);
-       venueService.getVenueById(request.getVenueId());
-        eventAttendeeRepository.removeEventAttendeeByEventId(eventId);
-        for(Integer attendeeId:request.getAttendeeId()){
+        for (Integer attendeeId : request.getAttendeeId()) {
             attendeeService.getAttendeeById(attendeeId);
+        }
+        Events existing = eventRepository.getAllEventByNameAndDate(request.getEventName(), request.getEventDate());
+        if (existing != null) {
+            throw new DuplicateException("Event with name " + request.getEventName() + " and date " + request.getEventDate() + " already exists");
+        }
+
+        for (Integer attendeeId : request.getAttendeeId()) {
+            attendeeService.getAttendeeById(attendeeId);
+        }
+
+        Events event = eventRepository.updateEvent(eventId, request);
+        eventAttendeeRepository.removeEventAttendeeByEventId(eventId);
+        for (Integer attendeeId : request.getAttendeeId()) {
             eventAttendeeRepository.saveEventAttendee(event.getEventId(), attendeeId);
         }
         return getEventById(event.getEventId());
+    }
+    @Override
+    public void deleteEventByEventId(Integer eventId) {
+       getEventById(eventId);
+       eventRepository.deleteEventById(eventId);
     }
